requirements.txt
# Bulletproof requirements for Streamlit Cloud
# Tested and guaranteed to work without installer errors
streamlit
google-generativeai
requests

# FastAPI Backend Dependencies
fastapi
uvicorn[standard]
pydantic

# Google Calendar API and OAuth dependencies
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# Security and encryption for token management
cryptography
python-dotenv

# Additional utilities
urllib3
